Changed the behaviour of Decimal class so that Decimal::rationalize() is no longer called after every operation / construction. This will reduce the volatility of the fractional precision (Decimal::m_places) of Decimals, and enable "stable rounding" to occur (digits are retained even if they become zeroes). This will speed the addition of large numbers of currency representations, in particular, where these are all stored as Decimals with the same number of places. This will also make it easier for users of the Decimal class to ensure that newly created Decimals can be safely added to existing ones, by rounding the newly created Decimals to the common number of places upon construction, and propagating an exception if this cannot be safely done.
I chose this over an alternative solution to this problem, which would have
involved storing a "minimum precision" as a class template parameter.
This would have been more complex and less flexible than approach described
above.

These changes had no affect on the unit tests (which were all passing),
except for the output operator's tests: all digits of precision are
printed, even if zeroes; and this is the desirable default behaviour; so
I changed these tests to make them pass.