Got rid of preprocessor option to disable arithmetic checking in the Decimal class.
With this option gone there are now fewer
"compilation paths" to worry about: the checking always occurs.
The implementation of this option cluttered the code and I had not
implemented it consistently anyway. It also added to the complexity of the
API.

I may need to revisit this issue if I ever reimplement the Decimal class with
an arbitrary precision integer as the underlying integral type. But if I do do
that, it would probably be better to turn off checking by means of some kind
of template specialization, than by means of a preprocessor directive.