Changed JEWEL_ASSERT to call std::terminate, rather than throwing an exception. This is because we want it to be more-or-less a drop-in replacement for assert. If we make it throw an exception, this may cause worries about exception-safety in places where it must be absolutely impossible for an exception to be thrown (e.g. destructors). Nevertheless by calling std::terminate rather than std::abort, we give the terminate-handler an opportunity to do something if it wants to.
Throughout most of the Jewel library replaced "assert" with "JEWEL_ASSERT".

In the implementation of jewel::Log, put mechanisms in place to ensure
that exceptions can never be thrown, even if the logging stream has
exceptions enabled: logging ceases if either exceptions are enabled on
the stream, or if an actual error state is detected on the stream.